requirements.txt
flask>=3.0.0
pypdf
python-docx
pandas
openpyxl
python-dotenv
requests
